+++ core/gethwrappers/generation/generate_link/json_utils/json_normalization.go
@@ -1,4 +1,4 @@
-package utils
+package json_utils
 
 import (
 	"bufio"
